requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
bcrypt==4.0.1
    # via paramiko
boto3==1.26.144
    # via
    #   -r requirements.in
    #   moto
botocore==1.29.144
    # via
    #   boto3
    #   moto
    #   s3transfer
build==0.10.0
    # via pip-tools
cachetools==5.3.1
    # via google-auth
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via
    #   cryptography
    #   pynacl
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.1.0
    # via requests
click==8.1.7
    # via pip-tools
coverage==7.6.4
    # via -r requirements.in
cryptography==41.0.0
    # via
    #   moto
    #   paramiko
delegator-py==0.1.1
    # via -r requirements.in
distlib==0.3.7
    # via virtualenv
filelock==3.12.3
    # via virtualenv
google-api-core[grpc]==2.11.0
    # via
    #   -r requirements.in
    #   google-api-python-client
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-storage
google-api-python-client==2.88.0
    # via -r requirements.in
google-auth==2.19.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-bigquery==3.10.0
    # via -r requirements.in
google-cloud-core==2.3.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-storage==2.9.0
    # via -r requirements.in
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.5.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.59.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.54.2
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   grpcio-status
grpcio-status==1.54.2
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
identify==2.5.27
    # via pre-commit
idna==3.4
    # via requests
jinja2==3.1.2
    # via
    #   -r requirements.in
    #   moto
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markupsafe==2.1.2
    # via
    #   -r requirements.in
    #   jinja2
    #   werkzeug
moto[s3]==4.2.0
    # via -r requirements.in
ndjson==0.3.1
    # via -r requirements.in
nodeenv==1.8.0
    # via pre-commit
numpy==1.26.0
    # via
    #   pandas
    #   pyarrow
oauth2client==4.1.3
    # via -r requirements.in
packaging==23.1
    # via
    #   build
    #   google-cloud-bigquery
pandas==2.1.1
    # via -r requirements.in
paramiko==3.2.0
    # via
    #   -r requirements.in
    #   pysftp
parse==1.19.0
    # via -r requirements.in
pexpect==4.8.0
    # via delegator-py
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==3.10.0
    # via virtualenv
pre-commit==3.3.3
    # via -r requirements.in
proto-plus==1.22.2
    # via google-cloud-bigquery
protobuf==4.23.2
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
ptyprocess==0.7.0
    # via pexpect
py-partiql-parser==0.3.6
    # via moto
pyarrow==13.0.0
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via
    #   google-auth
    #   oauth2client
pycparser==2.21
    # via cffi
pynacl==1.5.0
    # via paramiko
pyparsing==3.0.9
    # via httplib2
pyproject-hooks==1.0.0
    # via build
pysftp==0.2.9
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   google-cloud-bigquery
    #   moto
    #   pandas
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   moto
    #   pre-commit
    #   responses
requests==2.31.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   moto
    #   responses
responses==0.23.3
    # via moto
rsa==4.9
    # via
    #   google-auth
    #   oauth2client
s3transfer==0.6.1
    # via boto3
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   oauth2client
    #   python-dateutil
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
types-pyyaml==6.0.12.11
    # via responses
typing-extensions==4.7.1
    # via filelock
tzdata==2023.3
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.16
    # via
    #   botocore
    #   google-auth
    #   requests
    #   responses
virtualenv==20.24.3
    # via pre-commit
werkzeug==2.3.7
    # via moto
wheel==0.41.2
    # via pip-tools
xmltodict==0.13.0
    # via moto

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
